// src/modules/strategies/types/strategy.ts
export type Strategy = {
  id: number;
  strategy_name: string;
  strategy_type: string | null;
  trade_style: string | null;
  active: boolean;
  created_at: string;
  updated_at: string;
  deleted_at: string | null;
};

export type StrategyFormValues = {
  strategy_name: string;
  strategy_type: string;
  trade_style: string;
  active: boolean;
};

export type StrategyInsert = StrategyFormValues;

export type StrategyUpdate = Partial<StrategyFormValues> & {
  deleted_at?: string | null;
};